Library API + end-to-end design study; robustness fixes from adversarial V&V
src/ is now an importable package (__init__.py with lazy, side-effect-free
imports): from src.thrust.sk_models import cj_calc, ph_calc, axial_calc, ...

sk_models.py — pure calc functions extracted from the CLI stages (cj_calc /
ph_calc / axial_calc; stages now wrap them for JSON+print); NEW
axial_calc(chem='frozen') exposes the frozen-composition lower bound of the
product expansion (H2/air -6.8%, C2H4/O2 -10.0% on sonic T/Mdot, matching
the gamma_phase_audit section-4 bounds); portable mechanism ids in the JSON
records (kerosene entries previously stored an absolute build-tree path);
scoped silencing of the harmless Cantera ChemEquil fallback warning; clear
CLI usage/unknown-case messages; FileNotFoundError with remedies for
missing mechanism files.

stechmann_nozzle.py — stdout of 'validate' printed only 16 of the 18 rows
and dropped the '18/18 rows PASS' summary (bad hardcoded slice; file copy
was always complete); optimizer grids/tolerances generalized to any Pcp
(XTOL KeyError outside {20,200} atm before); _prop() guard with actionable
messages for unknown propellants and phi outside 0.2-3.0; package-relative
import of st_core with script-mode fallback.

cj_sweeps.py — phi block resumability was per-section (an interrupted run
skipped missing fuels on resume): now per-fuel. tables.py — wrapped in
main() so importing it is side-effect free. q_formal.py — plain-float dict
repr (no np.float64 noise under NumPy 2).

examples/example_design_study.py — NEW: end-to-end design study
(C2H4/O2, Rbar=45mm gap=5mm L=80mm, mdot=0.30 kg/s, Pcp=10 atm): CJ state,
Wolanski wave number (PCI 34 Eqs. 4-5) with honest l_cr band, SK
pressure-history + axial thrust, frozen bound, live eta_FJ, Stechmann
bell-vs-aerospike optimization; every step cross-checked against the
shipped validated JSONs. Physics of all validated modules untouched:
18/18, 8/8, 99/99, q_mapping 0.00% re-verified after the refactor.

+++ src/cycles/q_formal.py
@@ -147,7 +147,7 @@ def main():
         q_cj = sum((Y1.get(s, 0.0) - Ycj.get(s, 0.0)) * hf[s]
                    for s in set(Y1) | set(Ycj))
         Xcj = {s: x for s, x in zip(gcj.species_names, gcj.X) if x > 1e-4}
-        diss = {s: round(x, 5) for s, x in sorted(Xcj.items(), key=lambda kv: -kv[1])
+        diss = {s: float(round(x, 5)) for s, x in sorted(Xcj.items(), key=lambda kv: -kv[1])
                 if s not in majors}
         c.update(U_CJ=ucj, T_CJ=gcj.T, P_CJ_bar=gcj.P / 1e5,
                  q_CJplane_MJkg=q_cj / 1e6,
